+++ sentry-graphql/src/main/java/io/sentry/graphql/SentryInstrumentation.java
@@ -18,12 +18,14 @@ import graphql.schema.GraphQLNonNull;
 import graphql.schema.GraphQLObjectType;
 import graphql.schema.GraphQLOutputType;
 import io.sentry.Breadcrumb;
+import io.sentry.Hint;
 import io.sentry.IScopes;
 import io.sentry.ISpan;
 import io.sentry.NoOpScopes;
 import io.sentry.Sentry;
 import io.sentry.SentryIntegrationPackageStorage;
 import io.sentry.SpanStatus;
+import io.sentry.TypeCheckHint;
 import io.sentry.util.StringUtils;
 import java.util.ArrayList;
 import java.util.Arrays;
@@ -300,13 +302,16 @@ public final class SentryInstrumentation
     return environment -> {
       final @Nullable ExecutionStepInfo executionStepInfo = environment.getExecutionStepInfo();
       if (executionStepInfo != null) {
+        Hint hint = new Hint();
+        hint.set(TypeCheckHint.GRAPHQL_DATA_FETCHING_ENVIRONMENT, environment);
         scopesFromContext(parameters.getExecutionContext().getGraphQLContext())
             .addBreadcrumb(
                 Breadcrumb.graphqlDataFetcher(
                     StringUtils.toString(executionStepInfo.getPath()),
                     GraphqlStringUtils.fieldToString(executionStepInfo.getField()),
                     GraphqlStringUtils.typeToString(executionStepInfo.getType()),
-                    GraphqlStringUtils.objectTypeToString(executionStepInfo.getObjectType())));
+                    GraphqlStringUtils.objectTypeToString(executionStepInfo.getObjectType())),
+                hint);
       }
       final TracingState tracingState = parameters.getInstrumentationState();
       final ISpan transaction = tracingState.getTransaction();